flask/requirements.txt
Flask
SpeechRecognition
pyttsx3
python-dotenv
fireworks-ai
flask-cors